requirements.txt
Flask==2.0.1
praw==7.4.0
python-dotenv==0.19.0
vaderSentiment==3.3.2